backend-djangorest/api/src/pyproject.toml
[tool.poetry]
name = "balhom"
version = "1.4.0"
description = ""
authors = ["Fabbo"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
django-stubs = "^1.16.0"
django-configurations = "^2.4.1"
django-filter = "^23.1"
django-cors-headers = "^3.14.0"
drf-yasg = "^1.21.5"
djangorestframework-simplejwt = "^5.2.2"
PyJWT = "^2.6.0"
cryptography = "^40.0.2"
Pillow = "^9.5.0"
celery = "^5.2.7"
django-celery-results = "^2.5.0"
django-celery-beat = "^2.5.0"
redis = "^4.5.4"
eventlet = "^0.33.3"
PyYAML = "^6.0"
uritemplate = "^4.1.1"
psycopg2-binary = "^2.9.6"
coverage = "^7.2.3"
beautifulsoup4 = "^4.12.2"
docutils = "^0.19"
gunicorn = "^20.1.0"
pycryptodome = "^3.17"
django-minio-storage = "^0.5.0"
minio = "^7.1.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
